move initQuartz method to db module
