+++ azkaban-common/src/test/java/azkaban/test/Utils.java
@@ -38,14 +38,4 @@ public class Utils {
     return new DatabaseOperator(new QueryRunner(dataSource));
   }
 
-  public static DatabaseOperator initQuartzDB() throws Exception {
-    final AzkabanDataSource dataSource = new EmbeddedH2BasicDataSource();
-
-    final String sqlScriptsDir = new File("../azkaban-web-server/src/test/resources/")
-        .getCanonicalPath();
-
-    final DatabaseSetup setup = new DatabaseSetup(dataSource, sqlScriptsDir);
-    setup.updateDatabase();
-    return new DatabaseOperator(new QueryRunner(dataSource));
-  }
 }
