+++ azkaban-db/src/test/java/azkaban/db/AzDBTestUtility.java
@@ -16,7 +16,21 @@
  */
 package azkaban.db;
 
+import java.io.File;
+import org.apache.commons.dbutils.QueryRunner;
+
 public class AzDBTestUtility {
+
+  public static DatabaseOperator initQuartzDB() throws Exception {
+    final AzkabanDataSource dataSource = new EmbeddedH2BasicDataSource();
+
+    final String sqlScriptsDir = new File("../azkaban-web-server/src/test/resources/")
+        .getCanonicalPath();
+
+    final DatabaseSetup setup = new DatabaseSetup(dataSource, sqlScriptsDir);
+    setup.updateDatabase();
+    return new DatabaseOperator(new QueryRunner(dataSource));
+  }
 
   public static class EmbeddedH2BasicDataSource extends AzkabanDataSource {
