+++ azkaban-web-server/build.gradle
@@ -64,6 +64,7 @@ dependencies {
   generateRestli deps.restliTools
 
   testCompile project(':test')
+  testCompile project(path: ':azkaban-db', configuration: 'testOutput')
   testRuntime deps.h2
 }
